VIC-5883 Added spacing to gatewayWrapper tags (#4144)
- updated gateway wrapper tags to be more widely spaced
- changed engine references from 8bit to 16bit ints
- fixed some unrelated test script errors, and audio comments

+++ engine/aiComponent/behaviorComponent/behaviors/iCozmoBehavior_fwd.h
@@ -35,7 +35,7 @@ namespace ExternalInterface {
 
 namespace external_interface {
   class GatewayWrapper;
-  enum class GatewayWrapperTag : uint8_t;
+  enum class GatewayWrapperTag : uint16_t;
 }
   
 namespace RobotInterface {
